Add fixtures for end-to-end transfer tests
